# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Socle SIRH — Sorbonne-Hôtel",
...
    "version": "1.0.0"
  },
  "paths": {
...
    "/sql": {
      "post": {
        "tags": [
          "SQL des modules"
        ],
        "summary": "Exécuter une requête sur le schéma de votre module",
        "description": "Une seule instruction SQL, avec ses paramètres nommés.\n\nLes paramètres s'écrivent `:nom` et se passent dans `params`. Ne jamais\nconcaténer une valeur saisie par l'utilisateur dans la requête : c'est\nexactement ce que ces paramètres évitent.\n\n    {\n      \"query\": \"select * from candidats where statut = :statut order by nom\",\n      \"params\": { \"statut\": \"nouveau\" }\n    }\n\nLa recherche par motif fonctionne de la même façon :\n\n    {\n      \"query\": \"select * from candidats where nom ilike :motif\",\n      \"params\": { \"motif\": \"%mar%\" }\n    }\n\nUne instruction qui ne renvoie pas de lignes (`insert`, `update`, `delete`,\n`create table`) renvoie `rows: []` et le nombre de lignes touchées.",
        "operationId": "executer_sql_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequeteSql"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReponseSql"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "Jeton de groupe": []
          }
        ]
      }
    },
    "/sql/script": {
      "post": {
        "tags": [
          "SQL des modules"
        ],
        "summary": "Exécuter plusieurs instructions — pour appliquer votre schema.sql",
        "description": "Plusieurs instructions séparées par des points-virgules, sans paramètres.\n\nC'est l'endpoint qu'utilise la console SQL du socle pour appliquer le\nfichier `schema.sql` d'un module. Tout passe dans une seule transaction :\nsi une instruction échoue, aucune n'est appliquée.",
        "operationId": "executer_script_sql_script_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScriptSql"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReponseScriptSql"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "Jeton de groupe": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ReponseScriptSql": {
        "properties": {
          "statements": {
            "type": "integer",
            "title": "Statements"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "statements",
          "message"
        ],
        "title": "ReponseScriptSql"
      },
      "ReponseSql": {
        "properties": {
          "columns": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Columns"
          },
          "rows": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Rows"
          },
          "row_count": {
            "type": "integer",
            "title": "Row Count"
          },
          "truncated": {
            "type": "boolean",
            "title": "Truncated"
          }
        },
        "type": "object",
        "required": [
          "columns",
          "rows",
          "row_count",
          "truncated"
        ],
        "title": "ReponseSql"
      },
      "RequeteSql": {
        "properties": {
          "query": {
            "type": "string",
            "title": "Query"
          },
          "params": {
            "additionalProperties": true,
            "type": "object",
            "title": "Params",
            "default": {}
          }
        },
        "type": "object",
        "required": [
          "query"
        ],
        "title": "RequeteSql"
      },
...
      "ScriptSql": {
        "properties": {
          "script": {
            "type": "string",
            "title": "Script"
          }
        },
        "type": "object",
        "required": [
          "script"
        ],
        "title": "ScriptSql"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "Jeton de groupe": {
        "type": "http",
        "description": "Le jeton remis à votre groupe, en en-tête `Authorization: Bearer <jeton>`.",
        "scheme": "bearer"
      }
    }
  }
}